refactor(): Rename AbstractDatabaseHealthIndicator to DatabaseHealthIndicator
