+++ lib/health-indicators/database/mongoose.health.ts
@@ -2,13 +2,16 @@ import { Injectable, Optional } from '@nestjs/common';
 import { InjectConnection } from '@nestjs/mongoose';
 import { Connection } from 'mongoose';
 import { promiseTimeout } from '../../utils';
-import { AbstractDatabaseHealthIndicator } from '../abstract/abstract-database-health-indicator';
+import { DatabaseHealthIndicator } from './database-health-indicator';
 
 @Injectable()
-export class MongooseHealthIndicator extends AbstractDatabaseHealthIndicator {
+export class MongooseHealthIndicator extends DatabaseHealthIndicator {
   /**
    * Initializes the typeorm indicator
+   *
    * @param connection The typeorm connection of the application context
+   *
+   * @public
    */
   constructor(@Optional() @InjectConnection() readonly connection: Connection) {
     super(connection);
