+++ lib/health-indicators/index.ts
@@ -1,4 +1,6 @@
 export * from './typeorm/typeorm.health';
 export * from './dns/dns.health';
-export * from './mongoose/mongoose.health';
+export * from './database/mongoose.health';
+export * from './database/typeorm.health';
 export * from './microservice/microservice.health';
+export * from './health-indicator';
